Initial GitHub upload
This python code works on any version of Raspberry Pi as far as I have been able to test. This code will connect to a Unifi Controller, read data from the Wifi airports and find any client devices with the "notes" section filled and display those on the attached LCD (20x4) display. It will match the client device to the airport it is connected to, and can make "smart" assumptions on the strength of the connection to further locate the device with more precision.